Replace all remaining chars with unichars
Also changes several functions to accept const arrays (removing the need
for casting)

+++ ZXingObjC/qrcode/encoder/ZXBlockPair.h
@@ -19,8 +19,8 @@
 @property (nonatomic, assign, readonly) int8_t *dataBytes;
 @property (nonatomic, assign, readonly) int8_t *errorCorrectionBytes;
 @property (nonatomic, assign, readonly) int errorCorrectionLength;
-@property (nonatomic, assign, readonly) int length;
+@property (nonatomic, assign, readonly) NSUInteger length;
 
-- (id)initWithData:(int8_t *)data length:(unsigned int)length errorCorrection:(int8_t *)errorCorrection errorCorrectionLength:(unsigned int)errorCorrectionLength;
+- (id)initWithData:(int8_t *)data length:(NSUInteger)length errorCorrection:(int8_t *)errorCorrection errorCorrectionLength:(unsigned int)errorCorrectionLength;
 
 @end
